Restructure session code to be more modular
